backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
mysqlclient
captcha
